# static/openapi.yaml
openapi: 3.1.0
info:
  title: Ra Coding Agent API
  description: |
    REST API for ra-coding-agent session management.

    Provides two actions: create a session with an initial prompt, or continue the conversation using a session ID.
  version: 1.0.0
  contact:
    name: ra-coding-agent

servers:
  - url: http://localhost:3000
    description: Local development server

paths:
  /ra:
    post:
      operationId: createSessionAndPrompt
      summary: Create session and send first prompt
      description: |
        Creates a new AI coding agent session and sends the first prompt.
        Returns the session ID for use in subsequent /ra-reply calls.
      tags:
        - Sessions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RaRequest"
            example:
              prompt: "What files are in my project?"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RaResponse"
              example:
                id: "550e8400-e29b-41d4-a716-446655440000"
                response: "What files are in the current directory?

Based on my analysis, the current directory contains:

- api/
- examples/
- static/
- tests/"
        "400":
          description: Missing prompt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /ra-reply:
    post:
      operationId: continueSession
      summary: Continue an existing session
      description: |
        Send a continuation prompt to an existing session.
        Requires the session ID from a previous /ra call.
      tags:
        - Sessions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RaReplyRequest"
            example:
              id: "550e8400-e29b-41d4-a716-446655440000"
              prompt: "Continue with the next step"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RaResponse"
        "400":
          description: Missing id or prompt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Session not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /openapi.yaml:
    get:
      operationId: getOpenApiSpec
      summary: Get OpenAPI specification
      description: Returns this OpenAPI specification file
      tags:
        - Documentation
      responses:
        "200":
          description: OpenAPI YAML file
          content:
            application/x-yaml:
              schema:
                type: string

  /openapi.json:
    get:
      operationId: getOpenApiJson
      summary: Get OpenAPI specification as JSON
      description: Returns the OpenAPI specification in JSON format
      tags:
        - Documentation
      responses:
        "200":
          description: OpenAPI JSON file
          content:
            application/json:
              schema:
                type: object

  /health:
    get:
      operationId: healthCheck
      summary: Health check
      description: Returns the health status of the server and number of active sessions
      tags:
        - Health
      responses:
        "200":
          description: Server is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
              example:
                status: ok
                activeSessions: 0

components:
  schemas:
    RaRequest:
      type: object
      required:
        - prompt
      properties:
        prompt:
          type: string
          description: The prompt to send to the agent

    RaReplyRequest:
      type: object
      required:
        - id
        - prompt
      properties:
        id:
          type: string
          format: uuid
          description: Session ID to continue
        prompt:
          type: string
          description: The continuation prompt to send

    RaResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Session ID
        response:
          type: string
          description: The agent's text response to the prompt

    SessionEvent:
      type: object
      properties:
        type:
          type: string
          enum: [thinking, tool_call, tool_result, message, error, done]
          description: Event type
        data:
          type: object
          description: Event data

    MessageResponse:
      type: object
      properties:
        message:
          type: string
          description: Success message

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error description

    HealthResponse:
      type: object
      properties:
        status:
          type: string
          description: Health status
          enum: [ok]
        activeSessions:
          type: number
          description: Number of active sessions

  parameters:
    SessionId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Session identifier